+++ docs/hazmat/primitives/index.rst
@@ -14,6 +14,5 @@ Primitives
     mac/index
     cryptographic-hashes
     symmetric-encryption
-    smime
     padding
     twofactor
